Split up mocks.h
