+++ test/mocks/mesh.h
@@ -1,72 +1,16 @@
 #pragma once
 
-#include <cstddef>
 #include <cstdint>
-#include <cstring>
 #include <optional>
 
 #include "helpers/BaseChatMesh.h"
 #include "helpers/SimpleMeshTables.h"
+#include "radio.h"
+#include "random.h"
+#include "time.h"
 
-class FakeMillis final : public mesh::MillisecondClock {
-public:
-  unsigned long getMillis() override {
-    return 0;
-  }
-};
-
-class FakeRtc final : public mesh::RTCClock {
-public:
-  explicit FakeRtc(uint32_t initial_time) : _time(initial_time) {}
-
-  uint32_t getCurrentTime() override {
-    return _time;
-  }
-
-  void setCurrentTime(uint32_t time) override {
-    _time = time;
-  }
-
-private:
-  uint32_t _time;
-};
-
-class FakeRng final : public mesh::RNG {
-public:
-  void random(uint8_t* dest, size_t sz) override {
-    memset(dest, 0x5A, sz);
-  }
-};
-
-class FakeRadio final : public mesh::Radio {
-public:
-  int recvRaw(uint8_t*, int) override {
-    return 0;
-  }
-
-  uint32_t getEstAirtimeFor(int) override {
-    return 1;
-  }
-
-  float packetScore(float, int) override {
-    return 1.0f;
-  }
-
-  bool startSendRaw(const uint8_t*, int) override {
-    return true;
-  }
-
-  bool isSendComplete() override {
-    return true;
-  }
-
-  void onSendFinished() override {}
-
-  bool isInRecvMode() const override {
-    return true;
-  }
-};
-
+// No-op packet manager for native tests.
+// Satisfies Mesh dependencies while preventing packet allocation or queuing.
 class NoopPacketManager final : public mesh::PacketManager {
 public:
   mesh::Packet* allocNew() override {
@@ -108,6 +52,8 @@ public:
   }
 };
 
+// Test chat mesh for native tests.
+// Exposes packet receive handling and captures discovered contacts for assertions.
 class TestChatMesh final : public BaseChatMesh {
 public:
   TestChatMesh(mesh::Radio& radio, mesh::MillisecondClock& ms, mesh::RNG& rng, mesh::RTCClock& rtc,
@@ -156,6 +102,8 @@ protected:
   void onContactResponse(const ContactInfo&, const uint8_t*, uint8_t) override {}
 };
 
+// Test mesh context for native tests.
+// Owns mock dependencies in construction order and exposes the mesh via operator->.
 struct TestMeshContext {
   explicit TestMeshContext(uint32_t current_timestamp)
       : rtc(current_timestamp), mesh(radio, ms, rng, rtc, packet_manager, tables) {}
